+++ activerecord/test/cases/associations/has_many_associations_test.rb
@@ -1582,26 +1582,6 @@ class HasManyAssociationsTest < ActiveRecord::TestCase
     assert firm.companies.exists?(name: "child")
   end
 
-  def test_restrict_with_error_is_deprecated_using_key_many
-    I18n.backend = I18n::Backend::Simple.new
-    I18n.backend.store_translations :en, activerecord: { errors: { messages: { restrict_dependent_destroy: { many: "message for deprecated key" } } } }
-
-    firm = RestrictedWithErrorFirm.create!(name: "restrict")
-    firm.companies.create(name: "child")
-
-    assert !firm.companies.empty?
-
-    assert_deprecated { firm.destroy }
-
-    assert !firm.errors.empty?
-
-    assert_equal "message for deprecated key", firm.errors[:base].first
-    assert RestrictedWithErrorFirm.exists?(name: "restrict")
-    assert firm.companies.exists?(name: "child")
-  ensure
-    I18n.backend.reload!
-  end
-
   def test_restrict_with_error
     firm = RestrictedWithErrorFirm.create!(name: "restrict")
     firm.companies.create(name: "child")